+++ packages/react/src/exampleTypes/ResultPager.js
@@ -2,27 +2,31 @@ import _ from 'lodash/fp'
 import React from 'react'
 import { observer } from 'mobx-react'
 import InjectTreeNode from '../utils/injectTreeNode'
+import DefaultIcon from '../DefaultIcon'
 
 // These are to prevent warning from `active`, `previous`, `next`
 let span = ({ children }) => <span>{children}</span>
 let a = ({ children, onClick }) => <a onClick={onClick}>{children}</a>
 
 let ResultPager = InjectTreeNode(
-  observer(({ node, tree, List = 'div', Item = span, Link = a }) => {
+  observer(({ node, tree, Item = span, Link = a, Icon = DefaultIcon }) => {
     let pages = Math.ceil(
       (node.context.response.totalRecords || 1) / node.pageSize
     )
     let page = node.page || 1
     return (
-      <List>
+      <div className='contexture-result-pager'>
         <Item disabled={!(page > 1)}>
           <Link
             previous
             onClick={() => tree.mutate(node.path, { page: page - 1 })}
           >
-            {'←'}
+            <Icon icon='PreviousPage' />
           </Link>
         </Item>
+        {page > 3 && (<Item onClick={() =>
+          tree.mutate(node.path, { page: _.max([0, page - 5]) })
+        }><Icon icon="Previous5Pages" /></Item>)}
         {_.reverse(
           _.times(
             n =>
@@ -58,12 +62,15 @@ let ResultPager = InjectTreeNode(
             ),
           2
         )}
+        {page + 2 < pages && (<Item onClick={() =>
+          tree.mutate(node.path, { page: _.min([pages, page + 5]) })
+        }><Icon icon="Next5Pages" /></Item>)}
         <Item disabled={!(page < pages)}>
           <Link next onClick={() => tree.mutate(node.path, { page: page + 1 })}>
-            {'→'}
+            <Icon icon='NextPage' />
           </Link>
         </Item>
-      </List>
+      </div>
     )
   })
 )